Minor landscape fix.
Plus gamepiece jumping animation fix.

+++ Assets/Fourzy/Scripts/_Updates/UI/Helpers/OnRatio.cs
@@ -27,7 +27,7 @@ namespace Fourzy._Updates.UI.Helpers
 
         private float previousAspect;
 
-        private static float aspect => Screen.width / Screen.height;
+        private static float aspect => (float)Screen.width / Screen.height;
 
         protected void Start()
         {
